+++ sdk/dependencymap/azure-mgmt-dependencymap/generated_samples/maps_get_connections_for_process_on_focused_machine.py
@@ -47,6 +47,6 @@ def main():
     ).result()
 
 
-# x-ms-original-file: 2025-01-31-preview/Maps_GetConnectionsForProcessOnFocusedMachine.json
+# x-ms-original-file: 2025-05-01-preview/Maps_GetConnectionsForProcessOnFocusedMachine.json
 if __name__ == "__main__":
     main()