+++ sdk/dependencymap/azure-mgmt-dependencymap/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Release History
 
+## 1.0.0b2 (2025-05-09)
+
+### Features Added
+
+  - Model `ExportDependenciesRequest` added property `appliance_name`
+  - Method `ExportDependenciesRequest.__init__` has a new overload `def __init__(self: None, focused_machine_id: Optional[str], filters: Optional[_models.DependencyMapVisualizationFilter], appliance_name: Optional[str])`
+
 ## 1.0.0b1 (2025-04-20)
 
 ### Other Changes